+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Unreleased
 ----------
 
 * [Feature] Delay option
+* [Feature] Retry option
 
 0.3.0 (2017/03/09)
 ------------------